+++ pymilo/pymilo_param.py
@@ -13,7 +13,8 @@ import sklearn.dummy as dummy
 import sklearn.ensemble as ensemble
 import sklearn.pipeline as pipeline
 import sklearn.preprocessing as preprocessing
-from sklearn.cross_decomposition import PLSRegression, PLSCanonical, CCA
+import sklearn.cross_decomposition as cross_decomposition
+import sklearn.feature_extraction as feature_extraction
 
 quantile_regressor_support = False
 try:
@@ -85,7 +86,7 @@ except BaseException:
 OVERVIEW = """
 PyMilo is an open source Python package that provides a simple, efficient, and safe way for users to export pre-trained machine learning models in a transparent way.
 """
-PYMILO_VERSION = "1.2"
+PYMILO_VERSION = "1.3"
 NOT_SUPPORTED = "NOT_SUPPORTED"
 PYMILO_VERSION_DOES_NOT_EXIST = "Corrupted JSON file, `pymilo_version` doesn't exist in this file."
 UNEQUAL_PYMILO_VERSIONS = "warning: Installed PyMilo version differs from the PyMilo version used to create the JSON file."
@@ -246,10 +247,25 @@ SKLEARN_PREPROCESSING_TABLE = {
     "TargetEncoder": TargetEncoder if target_encoder_support else NOT_SUPPORTED,
 }
 
+SKLEARN_FEATURE_EXTRACTION_TABLE = {
+    # for raw data:
+    "DictVectorizer": feature_extraction.DictVectorizer,
+    "FeatureHasher": feature_extraction.FeatureHasher,
+
+    # for image data:
+    "PatchExtractor": feature_extraction.image.PatchExtractor,
+
+    # for text data:
+    "CountVectorizer": feature_extraction.text.CountVectorizer,
+    "HashingVectorizer": feature_extraction.text.HashingVectorizer,
+    "TfidfTransformer": feature_extraction.text.TfidfTransformer,
+    "TfidfVectorizer": feature_extraction.text.TfidfVectorizer,
+}
+
 SKLEARN_CROSS_DECOMPOSITION_TABLE = {
-    "PLSRegression": PLSRegression,
-    "PLSCanonical": PLSCanonical,
-    "CCA": CCA,
+    "PLSRegression": cross_decomposition.PLSRegression,
+    "PLSCanonical": cross_decomposition.PLSCanonical,
+    "CCA": cross_decomposition.CCA,
 }
 
 KEYS_NEED_PREPROCESSING_BEFORE_DESERIALIZATION = {